move action button logic to BasicPushTemplate
